+++ test/clojure_lsp/refactor/transform_test.clj
@@ -551,6 +551,41 @@
                            " (:require"
                            "  [some :refer [bar baz Dee foo]]))"
                            "   foo bar baz Dee")))
+    (testing "unsorted used refer with less max-line-length"
+      (test-clean-ns {:settings {:clean {:sort {:refer {:max-line-length 30}}}}}
+                     (code "(ns foo.bar"
+                           " (:require"
+                           "   [some :refer [Dee foo bar baz bla blowning]]))"
+                           "   foo bar baz Dee bla blowning")
+                     (code "(ns foo.bar"
+                           " (:require"
+                           "  [some :refer [bar baz bla"
+                           "                blowning Dee"
+                           "                foo]]))"
+                           "   foo bar baz Dee bla blowning"))
+      (test-clean-ns {:settings {:clean {:sort {:refer {:max-line-length 30}}}}}
+                     (code "(ns foo.bar"
+                           " (:require"
+                           "  [some :refer [bar baz bla"
+                           "                blowning Dee"
+                           "                foo]]))"
+                           "   foo bar baz Dee bla blowning")
+                     (code "(ns foo.bar"
+                           " (:require"
+                           "  [some :refer [bar baz bla"
+                           "                blowning Dee"
+                           "                foo]]))"
+                           "   foo bar baz Dee bla blowning")))
+    (testing "unsorted used refer with infinite max-line-length"
+      (test-clean-ns {:settings {:clean {:sort {:refer {:max-line-length 0}}}}}
+                     (code "(ns foo.bar"
+                           " (:require"
+                           "   [some :refer [Dee foo bar baz bla blowning]]))"
+                           "   foo bar baz Dee bla blowning")
+                     (code "(ns foo.bar"
+                           " (:require"
+                           "  [some :refer [bar baz bla blowning Dee foo]]))"
+                           "   foo bar baz Dee bla blowning")))
     (testing "unsorted used refer with sort disabled"
       (test-clean-ns {:settings {:clean {:sort {:refer false}}}}
                      (code "(ns foo.bar"