+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - General
   - Use lower-case for refer/import/require sorting. #560 #561
   - Avoid removing comments when sorting/cleaning namespace. #559
+  - Break lines when sorting refers along with then new `:clean :sort :refer :max-line-length` setting with a default of `80`. #562
   - Deprecate `lens-segregate-test-references` in favor to `:code-lens :segregate-test-references`
   - Check for a default `.cljfmt.edn` config file for cljfmt config settings if no `:cljfmt-config-path` was provided. #563